+++ js/core/moret/moretManager.js
@@ -33,6 +33,8 @@ class moretManager {
 		this.mesh_transformer.intersect_parents_lattice_secondary_modules();
 		
 		this.lattice_creator.create_lattices_secondary_modules_hex();
+		this.mesh_transformer.intersect_parents_lattice_secondary_modules_hex();
+
 		this.lattice_creator.remove_first_mpri_cell_secondary_modules();	
 		this.lattice_creator.remove_first_msec_cell_secondary_modules();	
 		
@@ -48,6 +50,8 @@ class moretManager {
 		this.mesh_transformer.intersect_parents_lattice_first_module();	
 		
 		this.lattice_creator.create_lattices_first_module_hex();
+		this.mesh_transformer.intersect_parents_lattice_first_module_hex();
+
 		this.lattice_creator.remove_first_mpri_cell_first_module();				
 		this.lattice_creator.remove_first_msec_cell_first_module();